can_packet_api
Define API for CAN packet.
