+++ tests/software_tests/can/_test_can_packet.py
